# internal/gen/server.gen.go
// SPDX-FileCopyrightText: 2026 The jma-openapi contributors
//
// SPDX-License-Identifier: MPL-2.0
//
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package gen

import (
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Liveness check
	// (GET /healthz)
	GetHealthz(w http.ResponseWriter, r *http.Request)
	// List area metadata
	// (GET /v1/areas)
	GetV1Areas(w http.ResponseWriter, r *http.Request, params GetV1AreasParams)
	// Get a single area metadata
	// (GET /v1/areas/{areaCode})
	GetV1AreasAreaCode(w http.ResponseWriter, r *http.Request, areaCode string)
	// Get forecast by office code
	// (GET /v1/forecasts/{officeCode})
	GetV1ForecastsOfficeCode(w http.ResponseWriter, r *http.Request, officeCode string)
	// Get forecast for a weather area by office code
	// (GET /v1/forecasts/{officeCode}/areas/{areaCode})
	GetV1ForecastsOfficeCodeAreasAreaCode(w http.ResponseWriter, r *http.Request, officeCode string, areaCode string)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Liveness check
// (GET /healthz)
func (_ Unimplemented) GetHealthz(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// List area metadata
// (GET /v1/areas)
func (_ Unimplemented) GetV1Areas(w http.ResponseWriter, r *http.Request, params GetV1AreasParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get a single area metadata
// (GET /v1/areas/{areaCode})
func (_ Unimplemented) GetV1AreasAreaCode(w http.ResponseWriter, r *http.Request, areaCode string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get forecast by office code
// (GET /v1/forecasts/{officeCode})
func (_ Unimplemented) GetV1ForecastsOfficeCode(w http.ResponseWriter, r *http.Request, officeCode string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get forecast for a weather area by office code
// (GET /v1/forecasts/{officeCode}/areas/{areaCode})
func (_ Unimplemented) GetV1ForecastsOfficeCodeAreasAreaCode(w http.ResponseWriter, r *http.Request, officeCode string, areaCode string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetHealthz operation middleware
func (siw *ServerInterfaceWrapper) GetHealthz(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealthz(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetV1Areas operation middleware
func (siw *ServerInterfaceWrapper) GetV1Areas(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetV1AreasParams

	// ------------- Optional query parameter "parent" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "parent", r.URL.Query(), &params.Parent, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "parent", Err: err})
		return
	}

	// ------------- Optional query parameter "name" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "name", r.URL.Query(), &params.Name, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "name", Err: err})
		return
	}

	// ------------- Optional query parameter "officeName" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "officeName", r.URL.Query(), &params.OfficeName, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "officeName", Err: err})
		return
	}

	// ------------- Optional query parameter "child" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "child", r.URL.Query(), &params.Child, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "child", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetV1Areas(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetV1AreasAreaCode operation middleware
func (siw *ServerInterfaceWrapper) GetV1AreasAreaCode(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "areaCode" -------------
	var areaCode string

	err = runtime.BindStyledParameterWithOptions("simple", "areaCode", chi.URLParam(r, "areaCode"), &areaCode, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "areaCode", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetV1AreasAreaCode(w, r, areaCode)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetV1ForecastsOfficeCode operation middleware
func (siw *ServerInterfaceWrapper) GetV1ForecastsOfficeCode(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "officeCode" -------------
	var officeCode string

	err = runtime.BindStyledParameterWithOptions("simple", "officeCode", chi.URLParam(r, "officeCode"), &officeCode, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "officeCode", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetV1ForecastsOfficeCode(w, r, officeCode)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetV1ForecastsOfficeCodeAreasAreaCode operation middleware
func (siw *ServerInterfaceWrapper) GetV1ForecastsOfficeCodeAreasAreaCode(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "officeCode" -------------
	var officeCode string

	err = runtime.BindStyledParameterWithOptions("simple", "officeCode", chi.URLParam(r, "officeCode"), &officeCode, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "officeCode", Err: err})
		return
	}

	// ------------- Path parameter "areaCode" -------------
	var areaCode string

	err = runtime.BindStyledParameterWithOptions("simple", "areaCode", chi.URLParam(r, "areaCode"), &areaCode, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "areaCode", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetV1ForecastsOfficeCodeAreasAreaCode(w, r, officeCode, areaCode)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/healthz", wrapper.GetHealthz)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/areas", wrapper.GetV1Areas)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/areas/{areaCode}", wrapper.GetV1AreasAreaCode)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/forecasts/{officeCode}", wrapper.GetV1ForecastsOfficeCode)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/forecasts/{officeCode}/areas/{areaCode}", wrapper.GetV1ForecastsOfficeCodeAreasAreaCode)
	})

	return r
}