# internal/gen/types.gen.go
// SPDX-FileCopyrightText: 2026 The jma-openapi contributors
//
// SPDX-License-Identifier: MPL-2.0
//
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package gen

import (
	"time"
)

// Area defines model for Area.
type Area struct {
	Children   []string `json:"children"`
	Code       string   `json:"code"`
	EnName     string   `json:"enName"`
	Name       string   `json:"name"`
	OfficeName string   `json:"officeName"`
	Parent     string   `json:"parent"`
}

// AreasResponse defines model for AreasResponse.
type AreasResponse struct {
	Items []Area `json:"items"`
}

// ErrorEnvelope defines model for ErrorEnvelope.
type ErrorEnvelope struct {
	Error ErrorModel `json:"error"`
}

// ErrorModel defines model for ErrorModel.
type ErrorModel struct {
	Code      string                  `json:"code"`
	Details   *map[string]interface{} `json:"details,omitempty"`
	Message   string                  `json:"message"`
	RequestId string                  `json:"requestId"`
}

// ForecastAreaResponse weatherArea 単位の個別予報レスポンス。temperature 系項目は含まない。
type ForecastAreaResponse struct {
	Office           OfficeRef   `json:"office"`
	PublishingOffice string      `json:"publishingOffice"`
	ReportDatetime   time.Time   `json:"reportDatetime"`
	WeatherArea      WeatherArea `json:"weatherArea"`
}

// ForecastResponse defines model for ForecastResponse.
type ForecastResponse struct {
	Office           OfficeRef         `json:"office"`
	PublishingOffice string            `json:"publishingOffice"`
	ReportDatetime   time.Time         `json:"reportDatetime"`
	TemperatureAreas []TemperatureArea `json:"temperatureAreas"`
	WeatherAreas     []WeatherArea     `json:"weatherAreas"`
}

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Service   string    `json:"service"`
	Status    string    `json:"status"`
	Timestamp time.Time `json:"timestamp"`
	Version   string    `json:"version"`
}

// OfficeRef defines model for OfficeRef.
type OfficeRef struct {
	Code string `json:"code"`
	Name string `json:"name"`
}

// TemperatureArea defines model for TemperatureArea.
type TemperatureArea struct {
	Code       string                       `json:"code"`
	Name       string                       `json:"name"`
	TimeSeries []TemperatureTimeSeriesEntry `json:"timeSeries"`
}

// TemperatureTimeSeriesEntry defines model for TemperatureTimeSeriesEntry.
type TemperatureTimeSeriesEntry struct {
	Temp *string   `json:"temp,omitempty"`
	Time time.Time `json:"time"`
}

// WeatherArea defines model for WeatherArea.
type WeatherArea struct {
	Code       string                   `json:"code"`
	Name       string                   `json:"name"`
	TimeSeries []WeatherTimeSeriesEntry `json:"timeSeries"`
}

// WeatherTimeSeriesEntry defines model for WeatherTimeSeriesEntry.
type WeatherTimeSeriesEntry struct {
	Pop         *string   `json:"pop,omitempty"`
	Time        time.Time `json:"time"`
	Wave        *string   `json:"wave,omitempty"`
	Weather     *string   `json:"weather,omitempty"`
	WeatherCode *string   `json:"weatherCode,omitempty"`
	Wind        *string   `json:"wind,omitempty"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse = ErrorEnvelope

// GetV1AreasParams defines parameters for GetV1Areas.
type GetV1AreasParams struct {
	// Parent offices.parent でのフィルタ。
	Parent *string `form:"parent,omitempty" json:"parent,omitempty"`

	// Name offices.name での完全一致フィルタ。
	Name *string `form:"name,omitempty" json:"name,omitempty"`

	// OfficeName offices.officeName での完全一致フィルタ。
	OfficeName *string `form:"officeName,omitempty" json:"officeName,omitempty"`

	// Child offices.children に含まれる child code での完全一致フィルタ。
	Child *string `form:"child,omitempty" json:"child,omitempty"`
}