# openapi/openapi.yaml
# SPDX-FileCopyrightText: 2026 The jma-openapi contributors
#
# SPDX-License-Identifier: MPL-2.0

openapi: 3.0.3
info:
  title: jma-openapi
  version: 0.1.0
  description: |
    気象庁防災情報 JSON を扱いやすい REST API として公開する非公式ラッパー API。
servers:
  - url: http://localhost:8080
    description: local
  - url: https://example.run.app
    description: cloud-run
tags:
  - name: health
  - name: areas
  - name: forecasts
paths:
  /healthz:
    get:
      tags: [health]
      operationId: getHealthz
      summary: Liveness check
      description: アプリケーション自身の liveness を返す。
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /v1/areas:
    get:
      tags: [areas]
      operationId: getV1Areas
      summary: List area metadata
      description: 気象庁の office 一覧を返す。`parent`、`name`、`officeName`、`child` は完全一致の AND 条件として扱い、0 件時は `200` と空の `items` を返す。
      parameters:
        - in: query
          name: parent
          schema:
            type: string
          required: false
          description: offices.parent でのフィルタ。
        - in: query
          name: name
          schema:
            type: string
          required: false
          description: offices.name での完全一致フィルタ。
        - in: query
          name: officeName
          schema:
            type: string
          required: false
          description: offices.officeName での完全一致フィルタ。
        - in: query
          name: child
          schema:
            type: string
          required: false
          description: offices.children に含まれる child code での完全一致フィルタ。
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AreasResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '502':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /v1/areas/{areaCode}:
    get:
      tags: [areas]
      operationId: getV1AreasAreaCode
      summary: Get a single area metadata
      description: 指定した office code の詳細を返す。
      parameters:
        - in: path
          name: areaCode
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Area'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '502':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /v1/forecasts/{officeCode}:
    get:
      tags: [forecasts]
      operationId: getV1ForecastsOfficeCode
      summary: Get forecast by office code
      description: 指定した office code の予報情報を返す。
      parameters:
        - in: path
          name: officeCode
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForecastResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '502':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /v1/forecasts/{officeCode}/areas/{areaCode}:
    get:
      tags: [forecasts]
      operationId: getV1ForecastsOfficeCodeAreasAreaCode
      summary: Get forecast for a weather area by office code
      description: 指定した office code の forecast から対象 weather area のみを返す。レスポンスには temperature 系項目を含めない。
      parameters:
        - in: path
          name: officeCode
          required: true
          schema:
            type: string
        - in: path
          name: areaCode
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForecastAreaResponse'
        '404':
          description: officeCode が不正、または areaCode に一致する weatherArea が存在しない。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
components:
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    HealthResponse:
      type: object
      required: [status, service, version, timestamp]
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
          example: jma-openapi
        version:
          type: string
          example: 0.1.0
        timestamp:
          type: string
          format: date-time
    AreasResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Area'
    Area:
      type: object
      required: [code, name, enName, officeName, parent, children]
      properties:
        code:
          type: string
        name:
          type: string
        enName:
          type: string
        officeName:
          type: string
        parent:
          type: string
        children:
          type: array
          items:
            type: string
    ForecastResponse:
      type: object
      required: [office, publishingOffice, reportDatetime, weatherAreas, temperatureAreas]
      properties:
        office:
          $ref: '#/components/schemas/OfficeRef'
        publishingOffice:
          type: string
        reportDatetime:
          type: string
          format: date-time
        weatherAreas:
          type: array
          items:
            $ref: '#/components/schemas/WeatherArea'
        temperatureAreas:
          type: array
          items:
            $ref: '#/components/schemas/TemperatureArea'
    ForecastAreaResponse:
      type: object
      required: [office, publishingOffice, reportDatetime, weatherArea]
      description: weatherArea 単位の個別予報レスポンス。temperature 系項目は含まない。
      properties:
        office:
          $ref: '#/components/schemas/OfficeRef'
        publishingOffice:
          type: string
        reportDatetime:
          type: string
          format: date-time
        weatherArea:
          $ref: '#/components/schemas/WeatherArea'
    OfficeRef:
      type: object
      required: [code, name]
      properties:
        code:
          type: string
        name:
          type: string
    WeatherArea:
      type: object
      required: [code, name, timeSeries]
      properties:
        code:
          type: string
        name:
          type: string
        timeSeries:
          type: array
          items:
            $ref: '#/components/schemas/WeatherTimeSeriesEntry'
    WeatherTimeSeriesEntry:
      type: object
      required: [time]
      properties:
        time:
          type: string
          format: date-time
        weatherCode:
          type: string
          nullable: true
        weather:
          type: string
          nullable: true
        wind:
          type: string
          nullable: true
        wave:
          type: string
          nullable: true
        pop:
          type: string
          nullable: true
    TemperatureArea:
      type: object
      required: [code, name, timeSeries]
      properties:
        code:
          type: string
        name:
          type: string
        timeSeries:
          type: array
          items:
            $ref: '#/components/schemas/TemperatureTimeSeriesEntry'
    TemperatureTimeSeriesEntry:
      type: object
      required: [time]
      properties:
        time:
          type: string
          format: date-time
        temp:
          type: string
          nullable: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorModel'
    ErrorModel:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true